Refactors and consolidates test suite
Removes redundant test files and consolidates test logic into relevant modules.
Moves the M4B splitting test from its own file to the text_to_speech tests.
Corrects module path in start tests.
Renames translation tests file.
