+++ tests/test_start.py
@@ -175,7 +175,7 @@ def test_main_pdf_synthesis(mock_exists, mock_terminal_size, runner):
         mock_pdf_synthesizer.synthesize.return_value = None  # Mock synthesize method
         mock_get_file_extension.return_value = ".pdf"
         # Now invoke the main command with the PDF file mocking the synthesizer run
-        module_path = "audify.text_to_speech"
+        module_path = "audify.start"
         with patch(f"{module_path}.requests.post") as mock_synthesize_kokoro:
             mock_synthesize_kokoro.return_value = MagicMock()
             mock_synthesize_kokoro.return_value.json.return_value = {
